Plugins: Chore: Renamed instrumentation middleware to metrics middleware
